Add Synchronized Selection. Issue #83.
Adds a Maya "SelectionChanged" event callback, because it's required for the implementation.
When a user selects a Maya node in the Outliner (or viewport), the node will be selected in the Solver UI.
The "Attribute" browser isn't quite perfectly synchronized because it needs to select Attributes as well as Maya nodes - for future work.