Skip Jacobian measurement evaluation for markers that cannot be affected.
Add 'getMarkerAttributeRelationship' function to query the (cached)
relationship from the Maya attributes on the markers - precomputed via
the 'maya.cmds.mmSolverAffects' plug-in command.

Change the type of error count for 'findErrorToParameterRelationship'
function, to be the number of markers, rather than the number of
all errors.

Change 'mapping' to 'list', because 'mapping' assumes a hash
map/dictionary, but we are not using a hash map.

Issue #157.