Add marker and attribute status column to Solver UI.
The 'status' is computed when validation or running a solver.

The logic for calling validation functions was changed slightly
to ensure non-"mmSolver" functions could be run (such as mmSolverAffects).

The "used" hint for the UI is stored in attributes on markers and the
Collection (via Auxiliary attributes).

Issue #27 and #157.