Fix non-unique UUIDs causing duplicate nodes to be added to Solver UI.
Fixes issue #165.

The most common occurrence is when the same rig is referenced into the
Maya scene more than once AND the rig was saved in Maya 2016+.

When saving in a Maya version that contains the UUID feature (introduced
in Maya 2016) the UUIDs are saved into the referenced file.
The UUIDs are then read from the referenced file and having the same
scene file loaded into Maya will cause duplicate UUIDs to exist.