screenspacerigbake is updated with latest fastbake lib
disabled viewport pause and used latest fastbake lib to bake attributes.
few constant variables are added.
License text is added to top of the script.
added new feature "Select in Outliner", which can select nodes in outliner based on selection in the UI.
Still one mel command exist in the tool I could not find an easy python equivalent command to break connections of a node.
#192