Remove zoom offset. Changed slider to value re-mapping.
The zoom offset was removed. Adding the zoom offset stopped
the Pan/Zoom tool from being able to drag in the viewport
to zoom in/out.

Zoom slider now directly changes the camera zoom.

The slider to value remapping has been simplified.

Moved some "tool.py" functions to "lib.py" because they were not
intended for users to run, and they should be "hidden away".

Changed the default slider range from "1-100" to "0-100", so that
"50" is the middle value.

Removed unneeded use of "kwargs", because it can be replaced with
keyword arguments on the function. (When the keyword arguments are
in the function definition users do not need to read the function
source code to know what values can be used with the function.)