Nodes to calculate camera transform and focal length from vanishing points.
Using vanishing points (parallel lines in an image), we can approximate
the camera translation and rotation, and lens focal length, from a
single image.

The user needs 2 vanishing points (4 lines) to calculate the focal length,
or just 1 vanishing point (2 lines) to calculate the rotation
(and rough translation) to the XZ plane.

We also include new mathematical structs and functions (independant
of Maya), to provide the vector and matrix maths.

Issue #121.

Change-Id: I14015df639339d9b9a33d730a6d07271d668dd78