Make the image aspect ratio easier for users to set.
Allow users to change the film back aspect with the image width/height.

The pixel aspect ratio also allows users of HDV or Anamorphic footage to
input a correct aspect ratio for the input image.

The calibration maths is very sensitive to the image aspect ratio and
will likely fail if it's wrong - even with good line data.

Issue #121.

Change-Id: I177ceb2abe2773d40248702523a0e550c07265c1