Fix order of arguments mmSolver._api.solverstandard._compile_single_frame()
I have incorrectly passed the arguments to the function creating the
"single frame solver" recipe. This created a silent error where the
"precomputed data" (a `dict`) was passed to the
`remove_unused_objects` argument which then asserted the wrong type.

I have added more type asserts to mitigate type of problem, in this
file.

Issue #232