Subdivide individual line segments to an arbitrary level.
This allows adding more detail in areas that are needed without
adding line segments where they are not needed.

The tool works on the current Marker or Line selection. If a line is
selected the line is uniformly subdivided. If markers are selected,
the line segments around the markers are subdivided.