Fix gamma correction of EXR images.
The gamma correction is approximated with a fast 'pow' (power) function,
but might be slightly less accurate.

This change returns MStatus::kFailure when a floating point data type is
being written to - Maya (seemingly) does not properly support writing
EXR or other floating point file types with the C++ API MImage class.

Resizing images is still using the Maya MImage class, and is therefore
buggy (see TODO in the code).