Add mmLineBestFit node - compute best fit line for a set of points.
This node is used to compute a line that is the best line through a set
of data points.

For example this node can be used to visualise a straight line through a
number of Markers. So if you place a bunch of markers on a line then add
all the 'worldMatrix[0]' nodes into this node, the line computed will
represent a perfectly straight line. If lens distortion is perfectly
corrected for the shot, the Marker points will perfectly line up with
the plate.