Fix build errors on Linux with Google Log (glog)
Disable use of glog on Linux, and enable on Windows.
This seems to be the only way to keep both operating systems happy.

It's not clear what is causing the linking issue with glog on Linux,
however the easiest thing to do is disable the use, as we don't really
need it at all.