Add "Set Camera Origin Frame" tool.
This tool is good for when there is no model to line your camera to. The
tool allows you to move a camera back to the origin on the current frame
while maintaining the relative positions of all (unlocked) bundles.

The camera is also automatically scaled to a semi-consistent scale. The
idea is that the scale can be adjusted later.

This tool can also be used to align two different camera solves for the
same image sequence. If there is at least 1 overlapping frame, this tool
can be used to align the cameras (and bundles).