Solve Attributes As Sub-Blocks
This allows lens distortion and focal lengths to be solved by only
solving those attributes as a "sub-problem" with the blocks of
attributes.

When lens distortion and focal lengths are solved as part of a large
number of attributes (such as with bundle translates) the focal length
and lens distortion results in very small changes for the reduction of
error and therefore the solver "decides" to only solve the other
attributes and leaves the lens distortion and focal length unaffected.