Do not use "Camera Frame Offset" field for Copy Tracks in mmSolver.
Backport change to v0.3.x.

Many users do not want to use the "Camera Frame Offset" in 3DEqualizer,
and the Copy Tracks tool assumes that users will use that feature.

If users need to use a different start frame from the loaded plate they
should use the Export Tracks feature (because it has a start
frame/offset feature). I expect this use case is rare, and the Copy
Tracks tool is optimized for the majority use case.