Improve result quality of mmCameraPoseFromPoints command.
The command only needs a minimum of 3 points to compute a camera pose
from a set of 3D points.

The main trick to this implementation is to re-use the existing camera
film back/focal length and to flip the 3D points in Z when given to
OpenMVG - because OpenMVG has a different coordinate system to Maya.