Use common variable to hide/show Scene Graph Mode widget.
The purpose is to allow a single point in the code that can be changed
easilly.

If a user preference created in the future, this makes it easier to
implement/reason with.

If a developer needs to enable the widget for testing, this constant
variable makes it easy to do so.