Fix Load Markers tool always adding to the active collection.
When Markers are created they will trigger an event - when loading
markers with the Load Markers tool, the event will automatically add the
markers to the 'active' collection, which is not expected by the user.

To stop this we block the "marker created" events from triggering when
creating Markers in the Load Markers tool.