Remove 'object relationships' feature from Solver UI
The feature is technically still available, but simply disabled from
users accessing it using the UI.

The "object relationships" feature is intended to compute what
attributes affect what nodes. The feature is intended to speed up solves
by only evaluating parts of the scene that really need it.
However in reality computing these relationships is extremely slow and
do not provide many benifits. The feature was disabled with a user
preference in v0.3.x after it was introduced, now we remove the feature
from the UI as well - so most users will not be able to use the feature
anymore.