Solver UI - Remove "validate solver" features
The Solver UI could automatically detect invalid solves before running
them and inform the user these were invalid. It also speeds up the
solver since it didn't need to worry about "trying" to solve something
that was known to fail.

However the downside to this is UI slowness and overall solver
slow-downs. Instead this feature was made redundant by adding more
checks into the solver and improving the stability so that solves didn't
break things as much.

This feature was locked behind a preference, but some older users would
still have this feature enabled (for some reason).

This change removes the feature and ensures that any user with the
preference turned on will no longer enable the feature. Nobody should be
using this feature - it's slower and worse in every use case.