Embed OpenMVG inside mmSolver project.
This change was (unfortunately) required to fix a crash caused when
running on Linux (CentOS 7.x) compiled with GCC 6.3.1, and trying to use
the robust camera pose estimation features of mmSolver. The code worked
fine on Windows, but on Linux with GCC 6.3.1 the code would segfault
consistently. The actualy bug is unknown, but compiling OpenMVG along
with mmSolver in the same build fixes the problem.

This changes the license requirements of the project and includes the
use of the Mozilla Public License v2.0 license, which is still
compatible with GNU LGPL v3 - so it's all fine.

The build scripts with OpenMVG were discarded and new build scripts were
written for mmSolver, to simplify everything.

The details of what has changed is in the file
"external/openMVG/README.mayaMatchMoveSolver.md".

This change also fixes the Ceres Solver CMake project name naming to
UpperCamelCase "Ceres", but the target name as lower case "ceres" -
confusing but important to be compatible with the Ceres library in the
wild (when not building with the auto-downloading FindCeres.cmake file
for mmSolver).