Include LDPK v2.8 archive to speed up build times.
This is a cached version of LDPK, from the following URL:
https://www.3dequalizer.com/user_daten/sections/tech_docs/archives/ldpk-2.8.tgz

The file is included inside mmSolver as convenience to developers and
the file hosters, at the cost of slowing down the Git fetch times for the
mayaMatchMoveSolver repository.

Please note that the provided archive is edited from the original to
just contain the source code needed - no tests, documentation,
images or pre-compiled binaries are included.