Disable "Attribute Blocks" for "standard" solver by default.
Although in some solves seem to benefit from solving in "attribute
blocks", most do not, and it seems to only slow down the overall solve.

See this commit for more details:
a14b5293609d64865a44aa517c1cc268d494dd53

Perhaps in the future we can only enable this feature if Focal Length or
Lens Distortion attributes are being solved - but we need to
conclusively prove that this produces better quality results, in less
time.