Support 'trival' Maya attribute connections in MM Scene Graph.
'Trival' maya attribute connections are connections to direct data
stored on a node, like a connection between translateX and translateY on
the same node. In this case the translateX and Y are sharing the same
value. This is different from 'complex' Maya attribute connections where
there is some computation being performed inside the node, and cannot be
easily replicated in MM Scene Graph.