Copy Markers Tool - Do not convert any selection to a Marker.
Converting a node to a Marker class will cause automatic attributes to
added, with the Marker class always assuming that the given node *is* an
actual Marker, and does not check at all.

This change fixes the tool to stop that from happening when the user has
selected the wrong input nodes.