Add Frei-Chen Edge Detection algorithm.
The Edge Detecton algorithm can be choosen at runtime, and adjusted with
the mmRenderGlobals node attribute.

Fixes the viewport background (and image plane) not being drawn when the
edge detection mode is used.

Also increases the sensitivity of the edge detection "threshold" value,
and the maximum values have been changed as a result.