Remove type check for Collection.
Although this code is helpful and *should* be left as is, it reveals a
circular dependency between the 'mmSolver._api.execute' and
'mmSolver._api.collection' modules.

This circular dependency causes an exception to be raised on Maya 2022
when running in Python 2.x mode (it's fine in Python 3.7).
Interestingly, the issue is also not raised on Maya 2020 which is Python
2.7 - so this seems to be specific to Maya 2022 running in Python 2 mode
on Linux and Windows.