Fix complex connection detectiuon for animCurve output.
Testing for an animCurve first, before testing for the type of
connection fixes the issue.

Without this fix, mmSolver would incorrectly consider an attribute with
an animCurve to be a "complex connection" (AKA an unsupported
connection), and would then fall back to the Maya DAG.

I've added more test cases that will hopefully catch this case in
future.