Add Display Layer support to MM Renderer
There are many changes in this commit, here is a summary.

Edge detection has been disabled and is not working in this commit, as
is 'wireframe alpha'.

The passes have been split up into 'begin', 'display layers' and 'end'.
The render opations in 'begin' and 'end' passes are fixed - they don't
change. The Display Layers adjust based on the number of display layers
in the scene.

Added a mmDepth shader, to offset the Z-depth channel for a render pass.

Added mmLayerMerge which is an "uber" shader intended to control all
blending of (display layer) render targets.

The 'SceneRender' class can now (optionally) render on a transparent
black using the 'm_background_style' attribute. This is very helpful for
rendering specific passes that are then composited on top of other
layers (as we are doing for the Display Layers).

'DisplayLayer's can be enabled/disabled, and have a name to dynamically
look-up nodes to render.

Custom attributes are now expected to be on Display Layer nodes in Maya.
These custom attribute values will be used for the renders.