Marker::getPosXY removes overscan factor by default
This means that calling code will be able to automatically use the
associated MarkerGroup node's attributes (such as the overscanX and Y)
to adjust the translateX/Y values.

This is important since the reprojection code already takes this
overscan into account, but if the 2D distorted Marker positions do not,
we get a incorrect alignment of reprojection and 2D position data.