MM Renderer - Move depth pass rendering logic to new class.
This places existing logic from the DisplayLayer into specific class
that will be used directly.

The advantage of this approach is to reduce the code and logic in
DisplayLayer and self contain the shader creation in a class that will
release the shader once it's 'destructed'.

The disadvantage of this approach is that hide logic in a class where
it's harder to understand.