MM Renderer - Use objectSet to cache displayLayer members.
This is intended to be a performance improvement so that we don't need
to walk the DG/DAG to find the members of a displayLayer with special
logic, at each render update - we do this only "once" in the
RenderOverride::setup() method.

This change does introduce an issue where the object list is not updated
when the objectSet members are updated - an area of research for the
future.