MM Renderer - Fix Layer depth ordering when rendering edges.
As a result all Layer merge modes other than Z-Depth have been
deprecated - it will probably be too difficult to maintain those layer
modes. Maybe down the road they'll be added - but not yet.

Fixes rendering image planes in the background pass using node type
assumptions and 'imageplane' in the draw classification string.

Added a new render target named 'kBackgroundColorTarget' and
'kBackgroundDepthTarget', to keep the background render to be used for
hold-out renders. Unfortunately this extra render target seems to be
needed for the layering system to work. It may be removed or refactored
in the future, if that's possible.

The layer merge variables are now named after what the layer represents,
rather than A, B, and C.