Build Script - Move mmscenegraph to rust sub-dir
The mmscenegraph library has been refactored and moved under the
./rust/src/ and ./rust/cppbind/ directories.

This change is to reduce the duplicated code when other libraries are
added to the project written in Rust with C++ bindings (which is
planned).

This change should be a "refactor" - no change in behaviour - however in
practice some minor changes will probably be visible in the build system
and scripts, such as the renaming of the CMake option
'MMSCENEGRAPH_BUILD_TESTS' is now 'MM_LIBRARY_BUILD_TESTS'.