mmSolverLibs - Add Rust/C++ bindings for core/lens/image.
This moves a lot of code from the ./src/mmSolver/ and ./rust/
to ./lib, which is now compiled as 'mmSolverLibs'.

Adds a new 'mmimage' library which can read and write EXR images, as
well as read EXR headers and custom attributes. This library is not yet
used by mmSolver, but will be used to query EXRs of useful metadata for
matchmove purposes (such as focal length, or focus distance
information). Test files have been added.

The ./src/mmSolver/core/ files (except 'reprojection.cpp') have also
been moved into the mmSolverLibs library, because those files do not
depend on any Maya specific API features and can be safely segregated.

The 'mmlens' sub-library (inside 'mmSolverLibs') has been created by
copying ./src/mmSolver/lens/ files and making edits to remove the public
use of the LDPK plug-ins. Some template functions are also used to
improve performance with multiple 2D coordinates being processed at
once. Test files have been added for this.

This removes the runtime link dependency of the 'dl' or 'dlfcn-win32'
dependencies as this removes the need for LDPK plug-ins.
Instead the LDPK lens distortion maths is included fromheaders and the
actual LDPK symbols are never exposed outside 'mmsolverlibs', and so the
LDPK linktime dependency is also removed.

Unfortunately due to a limitation of Rust's linking, we are only able to
link a single library that uses Rust/CXX bindings. To work around this
the 'mmSolverLibs' library has been added, which combines all the C++
and Rust bindings into a single static library then links it with the
mmSolver plug-in.
https://cxx.rs/build/other.html#linking-the-c-and-rust-together