Calculate lens distortion from 'identity' values
We do not need to generate an identity ST-Map for the values to be
undistorted/redistorted. Instead we can calculate the identity values
in-place to avoid unneeded memory.

To further improve performance we can calculate both undistortion and
re-distortion in the same loop and output to a 4-channel RGBA 32-bit
pixel value. OpenEXR often(in OpenImageIO at least, I think?) reads full
RGBA 32-bit pixels anyway, even if only R and G are needed, so why not
save an extra image and pack both undistortion and re-distortion values
into a single image? This is what the "apply_undistort_and_redistort_*"
functions will do.