Fix incorrect counting in SolverObjectCountResult
The SolverObjectCountResult struct, used to keep track of the number of
objects to be solved was incorrectly keeping track of the number of
parameters, because of an "off-by-one" error.

When 1 parameter was solved with 3 frames, the initial
SolverObjectCountResult struct was constructed with an initial count of
1. When the 'SolverObjectCountResult.divide()' method is run the count
of '4' (3 + 1) was divided by '3' (3 frames with one parameter each),
with the result being '0'.

The issue is fixed by making the initial count of the
SolverObjectCountResult struct be 0, not 1.

A similar issue can be seen in the other structs: SolveValuesResult,
TimerResult and SolverResult.

Additionally this commit moves some complicated functions into 'fill'
methods so that the setting, adding, dividing and appendToMStringArray
logic is in one struct.

The 'test_solve_robotArm.py' test has been added to catch this issue
from happening again.

Issue #248.