Fix Solver Test 3 - Use larger delta for Maya evaluation
This test (test_solver/test3.py) was a little flakey because when using
the cminpack_lmder solver and evaluated with Maya DAG, the correct
values would not solve. This seems to be caused by using a very small
delta value. In the specific evaluation, the small delta value was not
picking up the changes (because it could not produce Differentiable
solves).