Add OpenGL Silhouette rendering to MM Renderer
Silhoutte rendering can be used to see the "edges" of a mesh
object/character so it's much easier to match a mesh to a plate.

This implementation of Silhouette uses direct raw OpenGL calls from the
MGLFunctionTable (Maya OpenGL Function Table), but unfortunately the
MGLFunctionTable is deprecated, so in the future this implementation may
need to be re-implemented.