Fix compilation of mmsolverlibs on Linux/GCC.
On Linux the order of the libraries given to the (GCC) linker matters a
lot.

This issue comes from the fact that mmsolverlibs_rust and mmsolverlibs
(C++) both use functions from each other and so there is a cyclic
dependency between them.

The transitive dependency that mmsolverlibs has on mmsolverlibs_rust
does not allow the rust dependency to be linked on Linux - on Windows
with MSVC, it's fine.

This page contained the critical information to help understand the
issue:
https://stackoverflow.com/questions/45135/why-does-the-order-in-which-libraries-are-linked-sometimes-cause-errors-in-gcc

Due to this change the test binaries for mmsolverlib has been re-enabled
(because they now work correctly).