Add GPU/CPU capacity and eviction to ImageCache.
This allows allocating CPU and GPU memory, and automatic eviction
(deallocation) of cached GPU/CPU memory, to keep a spesified number of
images in the cache.

The capacity is currently based on the number of images, not the amount
of memory - this will be changed in the future.

Tests show that the number of GPU images stored can be quite low, and we
can upload to the GPU from the CPU cache fairly fast (50 fps).

The next steps would be to pre-allocate the anticipated GPU images that
will be loaded in the future and allocate them from a secondary thread
(or concurrently) to ensure smooth playback. This would allow a small
number of GPU images backed by a much larger number of cached CPU
images.

GitHub Issue #252.