Respects Image Cache CPU/GPU memory capacity.
We add a "capacity bytes" getter/setter for the GPU and CPU.

We also keep track of how much memory is currently used inside the
cache, not just the total capacity.

The MTexture pointer is now wrapped in a struct, CacheTextureData, so we
can calculate the amount of memory that was used, and dimensions of the
texture. CacheTextureData takes control of it's own GPU memory
allocation, and deallocation using the underlying MTextureManager
pointer.

GitHub issue #252.