mmReadImage Command - Query channels and byte count
This allows the user to query the total amount of memory required by an
image in memory. In turn this allows us to make better decisions about
the Image Cache utilisation in Python (or MEL) code.

This also fixes inconsistencies with variable naming, and getting the
number of channels in an image. We now make assumptions about the data
type that will be read by the Maya MImage class.

GitHub issue #252.