Maya 2025 on Linux - Initial support added for building plug-in.
Uses OpenColorIO 2.3.2 (upgraded from 2.2 in previous versions).

This has not been tested well and has been confirmed to build, but there
are currently symbols missing at runtime (which is a problem currently
affecting all Linux builds with OpenColorIO).

The location of Qt's Resource Compiler executable ("rcc") and UI
Compiler ("uic") has been moved in the Linux install of Maya 2025 to the
'${MAYA_LOCATION}/libexec' directory. The CMake script has been updated
for this.