Upgrade CXX to 1.0.124
CXX is used to generate Rust/C++ bindings. The same version of CXX must
be used across the project.

We are updating just to stay update and not fall behind. There are no
expected features that are needed from this update. I am hoping this
will resolve linking/building issues but initial tests seem to suggest
that won't happen.

This is the latest version (1.0.124) as of today (2024-06-20).