Image Cache - Set capacity at mmSolver startup
This commit reorganises a lot of functions and modules of the image
cache.

When starting mmSolver (usually at the start of Maya), the Image Cache
capacity will be set from the preferences stored by the user.

The capacity preferences can be overridden per-maya-scene as needed too,
without changing the default for new Maya scenes.