Silhouette Renderer - Fix background colours during playblast
This fixes a bug where the viewport background was being set incorrectly
while Maya playblasts.

This seems to have been tracked down to the "MRenderer::useGradient()"
method which was reporting the wrong values during a playblast.

'setOverridesColors' was also set to override colours. This combination
of values appears to provide the correct output, although I'm not
exactly sure why or how.