Fix MM ImagePlane VP2 display filters
The display filters for the MM image plane shapes were not working
because both shared the same values and it seems Maya doesn't like that.

To avoid confusion the v1 MM image plane doesn't register a display
filter at all. This could be enabled with a simple value change if
needed, but hopefully nobody uses the old v1 image plane.

GitHub issue #263.