Build System - Build packages by default.
You need to explicitly disable the packages if you don't want to build
them.

I know that's annoying when most of the time you don't want to produce a
package, you just want to install locally, test, and iterate on the
code.

This change makes it possible that a default build from a random user
will produce a package ready for installation, which is probably what
they want.