Change Python package update system
The versions are hard-coded but also relaxed in some cases to allow
dependencies to be resolved easier.

The reason we hard-code version numbers is to ensure consistent
formatting/linting and documentation building.

Different Maya versions will require different Python versions and so
will require different package versions.

The "doc" and "dev" requirement files have been merged into a single
file for each Maya version because keeping all the files up to date
seemed like a bit of a nightmare, and doesn't seem to provide any extra
details. Instead the different needs of the packages are named in the
comments of the files.

The versions (where given) in these files are known to work today
(2024-09-16) with the Python versions expected. If the Python vesions
change in the future, these version numbers may need to be changed.