MM ImageCache - Update Python code for Python 2.7
We must make sure that byte size values are stored as 'long's in 2.7 and
as 'int' in 3.x.

We must also be careful of division, as Python 2.x does not convert
'1/2' to a floating point number (0.5), but Python 3.x does.

Finally we fix the circular dependencies of the initalization of the
image cache, so it's moved to a different module to avoid those
problems... and because it really is an entry point of sorts.