MM Image Cache - Fix erasing group items
When a user clears all the memory in a shot, the underlying
ImageCache::m_cpu_group_map would be changed, but the code is
iteratiting over it, so we must manually check the group exists each
time, and remove the first value each time. This probably has a
performance impact, but it's required for stability and safety.