Add Hi-DPI icons for mmSolver
These Hi-DPI icons follow the naming convention defined by Maya;
https://help.autodesk.com/view/MAYAUL/2025/ENU/?guid=GUID-F2900709-59D3-4E67-A217-4FECC84053BE
with the page name "Use a high resolution custom icon for a shelf item".

This includes a MS Windows Batch script to auto-generate all the icons
from the .svg files.

The README.md file is intended as documentation to help me remind myself
of the details if of the icons... and new people on the project.