Combine Rust workspace dependencies into single file.
This allows us to place the list of dependencies in the main workspace
file and then reference the individual packages in the respective
sub-crate. The advantage is to allow us to avoid duplicating and using
different versions of the same package in different sub-crates of the
workspace, which can increase compilation times and bloat the final
binaries.