Fix EXR reader memory leak
This change fixes a memory leak when reading EXR images.
The image reader also improves the speed of reading EXRs and can
optionally vertically flip the image data being read.

We also reuse allocated memory when reading an EXR image sequence. This
avoids memory fragmentation, but does waste a little bit of memory - in
the future we could clean this up and clear the temporary memory used -
if needed.