mmMemoryGPU - don't print verbose error from missing GPU
When a MRenderer cannot be found, lets just print that and allow the
command to continue successfully.

An MRenderer is probably not found if Maya is running in a batch-mode
without any GPU, for example in a Docker container, or the Linux machine
is not running a X11 display server.