Fix corrupted marker/attribute string values
From investigation, it seems that the issue probably stems from the
change to GCC 11.2 used on Linux only, or perhaps an internal change to
the Maya API, allowing "return value optimisation" or something like
that.

The issue was that the internal data inside the MString returned from
the 'Marker::getLongNodeName()' and 'Attr::getLongName()' method was
deallocated, but the object was on the stack, with a dangling pointer.
When `MString::asChar()` was called, the pointer pointed to deallocated
"random" memory which interpeted the "random" memory as bad characters.

The issue was fixed by:
1) Creating temporary MString variable that would hold the data until we
could call '.asChar()'.
2) Slightly adjust the ' Marker::getLongNodeName()' to remove unneeded
names (probably has no real affect on the code, just some clean up).

GitHub issue #273.