Fix potential dangling pointer usages.
Based on the fix for #273 (commit
667844c71d047b872d0eead8fd7e72203f92b401), this commit attempts to fix
the same pattern in all of the mmSolver code (with MString at least).

The issue is that the MString may be deallocated before the `.asChar()`
can be called, so the underlying string that is returned is corrupted.
This change is intended to fix that problem.

GitHub issue #273.