Maya Load Marker - Fix UI preferences bug
This change is intended to fix a bug where the Load Bundle Position
checkbox did not get the correct value.

We also reduced the number of duplicated strings for setting/getting
config values. There was a typo in the load_bundle_positions string - so
that also contributed to the bug.