Core Solver - Add Ceres LMDIF solver.
Add the Ceres library to bake a "LMDIF" solver, which will perform
Levenberg-Marquardt using "finite differences".

This change also requires the "solver verison" and "solver_type" to be
exposed to the Python API, so that the tests can run using Ceres.

Right nwo Ceres is similar to the "cminpack_lmdif" in terms of
performance, and does not yet match "cminpack_lmder". As a result, the
new "ceres_lmdif" will not be made default until the solver can be
proven to provide better results and integrates nicely, just like
"cminpack_lm*" solver types.

GitHub issue #174.