Core Solver - Remove "behind camera error"
There was a "behind camera error" value that was added to markers if the
bundle went behind the camera.

This error was only applied in the Maya DAG evaluation, not in MM Scene
Graph.

Running tests seems to suggest this error metric is not important to
overall solve stability, and should therefore be ignored.