Core Solver - Pre-compute Valid Frames
The frames that can or cannot be solved are validated and culled before
a proper solve can be started.

This is intended to avoid invalid solves from attempting to be solved,
and overall produce a more robust solver.

We also move the logic of whether the solver is valid from the Python
"actionstate.py" file to the C++ core, which gives us more information
to make better and more informed decisions about solver validity.

This is partially related to gitHub issue #174.