Add More Build Info to Assertions/Panics/Checks/TODOs.
The formatting and reorganisation of the assert code is alsso added, and
a Maya mmSolver layer is added similar to the debugging print statements
that were added.

We don't need to use static chars anymore, we can use "std::ostream"
too, which is nicer to add important details to assertion messages.

The build information must be given by the build script, this is
intentional because trying to get this information out of CMake is like
pulling teeth.