Core - Add Build Constants into .cpp files
This change reorganises the assert code, so we can more easily redirect
the output to another std::ostream, similar to what we do with the Maya
debug prints.

This change also reduces the number of warnings printed during
compilation due to redefined '#define's.

The intention of adding the build constants into .cpp files, rather than
.h files is that the build date/time will always change each run of the
build, which causes a number of files to be recompiled (because the
change is in the header file). By adding these values into the .cpp file
we can only recompile 2 files (1 for each mmsolverlibs and mmSolver).