Core Solver - Use time to evaluate marker/bundle affects relationships
This is (unfortunately) a very large change, but necessary. These
changes are made specifically because of the 'Ceres' solver support
required.

The change is included to improve stability and correctness of the
solver, avoid edge cases where the solver can incorrect evaluate invalid
objects.

This changes the fundamental data structures that the core solver uses
to store Markers, Attributes and Frames - rather than a simple list of
shared pointers, we include an 'enabled' state. This allows us to track
the validity of each object. These data structures are MarkerList,
AttrList and FrameList. The biggest difference for FrameList is the use
of FrameList.size() rather than FrameList.length().

We do not shrink/grow the respective vectors, because there are numerous
data structures and indices that all assume the index of these lists
does not change.

I've also adopted the MarkerIndex, AttrIndex, FrameIndex, and Count32
data type aliases throughout the code, using raw loops. These data types
are not "Strong Types", but are intended as a starting point to create
Strong Types - to catch problems at compile time, not runtime.

Frame numbers stored as as Maya 'MTime' has also been changed to using
simple integers. The logic is to avoid using Maya types as much as
possible.

We also use the MMSOLVER_ASSERT macro, rather than the C++ "assert"
macro, because there is additional logic used with MMSOLVER_ASSERT,
where these *are* run in release builds!

The change also includes a lot of fixes, and changes to the integration
tests. Changes to tests include:
- To include the use of frames when evaluating the 'affects'
relationships.
- Fixes to 'test_print_statistics.py'.
- Fixes to 'test_deviation_calculation.py' for 'numberOfParameters=8'
which should always have been there, but mmSolver did not compute
correctly.
- Introduction of 'frames' and many more tests for 'test_nodeaffects.py'
- Commented out many 'ceres_lmdif' and 'maya_dag' tests because the
solver produces bad results and only slows-down the test suite.
- Disable the 'test_camera_solve_garage.py' test, because it takes too
long to complete (~53 mins).

src/mmSolver/adjust/adjust_relationships.cpp/.h:
There are a LOT of changes in this file, including function name changes
to (hopefully) more understandable names.

src/mmSolver/adjust/adjust_base.cpp:
This is pretty much the core of the logic of how the solver is run, and
has been improved to add validity checks to avoid errors.

src/mmSolver/adjust/adjust_consoleLogging.cpp/.h:
The console logging formatting and logic has been centralised and
improved. Note that we now print more information by default compared to
mmSolver v0.5.2.

The 'removeUnusedFrames' flag has been added to the 'mmSolver' and
'mmSolver_v2' commands. This mirrors the 'removeUnusedMarkers' and
'removeUnusedAttributes' flags, but for frames.