Build System - Add MMSOLVER_USE_GLOG option
Enable or disable the use of glog.

This commit drops support for 'glog' on Linux. This means 'glog' is no
longer a dependency of mmSolver on Linux.

The reason we remove 'glog' on Linux is to avoid build errors, and to
fully use the embedded 'miniglog' implementation inside Ceres.

Before this change 'glog' was included on Linux simply for the use of
'glog/logging.h' because miniglog's 'glog/logging.h' was not found at
include time and we mistakenly included 'glog' to account for this.