3DEqualizer Tool - Copy/Export 2D Points with world-space transform data
This adds a version 5 ("v5") uvtrack format, that supports embedding 3D
transform matrix data for scene and point group data.

Improves the logic for getting camera/point groups and points in the
Copy and Export tools. The messages shown to the user are also less
direct and easier to understand.

I've also improved the opening/writing and closing of file handles,
using 'with' statements when possible.

I've improved the code generation of the "scriptdb" files, to improve
the iteration speed when writing changes to the "python/*.py" files.

GitHub issue #275.