Maya Tool - Fix Center 2D Offset node update bug
The Center 2D Offset UI was not updating the nodes used under the hood
when the centering was created after the UI was opened - it only found
the nodes when the UI was opened.

Added a "Camera" label, so the user can see which camera is being
controlled.

Renamed "Reset All" button to "Reset Offset" because the later is more
descriptive to what the button actually does - What does "Reset All"
mean? What is "all"?